deepen: Summarize module by inlining historical-context
Inline lib/ai/historical-context.ts (119 lines) as private helpers
inside lib/ai/summarize.ts, parallel to ADR-0002. The module exposed
two functions (getHistoricalSummaries + buildContextEnrichedPrompt)
and an interface (HistoricalSummary). It had exactly one production
caller — the historical-context enrichment block in summarize.ts —
and a 229-line unit test file. Classic extracted-for-testability shape.

Applying the deletion test from LANGUAGE.md: the module's interface
(two functions + one type + two constants) was nearly as complex as
its implementation (60 lines of body). Inlining concentrates the
historical-context logic where it actually runs. The shallow seam
between the helpers and their only caller carried no leverage —
callers couldn't swap it, no two adapters existed, and the test file
only covered the same surface that the surrounding summarize
pipeline already exercises end-to-end.

What this deepens: the Summarize module. Its interface is unchanged
(callers still call summarizeFiling). The internal implementation now
owns the historical-context logic instead of routing through a
sibling module that no one else used. Leverage: future changes to
historical-context behaviour live in one file. Locality: the
fetch-and-fold logic is collocated with the only place it executes.

Dependency category per process/Deepening/deepening.md: in-process
(Postgres via the existing getPrismaClient seam). No adapter needed —
the database access pattern matches summarize.ts's existing usage.

Tests deleted:
- __tests__/ai/historical-context.test.ts (229 lines exercising the
  shallow seam — covered now by summarize.ts's existing tests, which
  cross the deepened module's interface)

Tests adjusted:
- __tests__/ai/summarize-enrichment-gate.test.ts: removed the dead
  jest.mock('@/lib/ai/historical-context', ...) stanza; added
  summary.findMany to the prisma mock so the now-private helper has
  a stand-in. Same 7 assertions still pass.

CONTEXT.md: added "Historical Context" term that names the enrichment
pattern, with explicit note that it lives as private helpers inside
the Summarize module rather than as a standalone module.

ADRs respected:
- ADR-0001 (filing-type registry): unrelated.
- ADR-0002 (inline analysis-depth): direct precedent — same shape.
- ADR-0003 (extractor-registry withdrawal): unrelated.
- ADR-0004 (Lifetime Seat): unrelated.

No new ADR needed.
